tapset-utrace.cxx:check used PID in process probes
tapset-utrace: made a change so scripts with process(PID).*
and -x PID2, only validates PID (the relevant pid) and not PID2

+++ tapset-utrace.cxx
@@ -684,13 +684,13 @@ struct utrace_builder: public derived_probe_builder
     else if (has_null_param (parameters, TOK_END))
       flags = UDPF_END;
 
-    // Check that if a pid was given (through -x flag), then it corresponds to a running process.
+    // Check that if a pid was given, then it corresponds to a running process.
     if (has_pid || sess.target_pid)
       {
         int rc = -1;
         if (has_pid && pid > 0)
           rc = kill(pid, 0);
-        if (sess.target_pid > 0) // pid given using -x flag
+        else if (!has_pid && sess.target_pid > 0) // pid given using -x flag
           rc = kill(sess.target_pid, 0);
         if (rc == -1)
           switch (errno) // ignore EINVAL: invalid signal
